python-searcher/pyproject.toml
[tool.poetry]
name = "python-searcher"
version = "0.1.0"
description = ""
authors = ["Lucas B <lucas@jito.wtf>"]
readme = "README.md"
packages = [{ include = "python_searcher" }]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.51.1"
protobuf = "^4.21.12"
solders = "^0.14.2"
click = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-grpc-plugin]
proto_path = "../jito_protos/protos"
python_out = "./python_searcher/generated"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
